adding init files and renaming supabase to database to avoid conflict with pkg
